Livio-main/src/backend/requirements.txt
anyio==4.11.0
asgiref==3.8.1
attrs==25.4.0
auth0-api-python==1.0.0b5
Authlib==1.6.5
boto3==1.40.75
botocore==1.40.75
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
ecdsa==0.19.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python3-openid==3.2.0
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
thefuzz==0.22.1
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.44.0
whitenoise==6.8.2
pymysql==1.1.0